refactor: update drawer and navbar
